Split scrollable.tests.js into several files (T739041) (#8859)
* Split scrollable.tests.js to several files (T739041)

* Move markup and global variables settings to the root tests file

* Split root markup into each module.beforeEach function

(cherry picked from commit 077b1568f30af184caa51a1d555a43fc7823a9f2)
